refactor supabase plugin package to pare down main supabase.go file
